runtime: Do not gate `MockScheduler` behind feature
