+++ crates/grpc/Cargo.toml
@@ -21,7 +21,7 @@ tonic = { version = "0.8.3", default-features = false, features = ["codegen", "p
 tracing = { version = "0.1.37", features = ["attributes"] }
 
 tardigrade = { version = "0.1.0", path = "../.." }
-tardigrade-rt = { version = "0.1.0", path = "../rt", features = ["tokio", "test"], optional = true }
+tardigrade-rt = { version = "0.1.0", path = "../rt", features = ["tokio"], optional = true }
 tardigrade-worker = { version = "0.1.0", path = "../worker" }
 
 [dev-dependencies]
@@ -30,6 +30,8 @@ assert_matches = "1.5.0"
 tokio = { version = "1.24.1", features = ["rt", "macros"] }
 tonic = "0.8.3"
 version-sync = "0.9.4"
+
+tardigrade-rt = { version = "0.1.0", path = "../rt", features = ["test"] }
 
 [build-dependencies]
 tonic-build = "0.8.4"
